refactor(observer): consolidate duplicated session-command, provider, and test wiring
Three independent, behavior-preserving deduplications:

- Session commands: hoist lookupWorktree, validateSnapshotRow,
  commandValidationError, and a unified buildEnsureAgentWorkspaceIntent into
  session/shared.ts; create/startAgent/resumeAgent now reuse them.
- CLI provider registry: extract applyObserverPaths and applyHarnessAgentOptions
  in observerProviders.ts, collapsing the per-harness option-building if-chain.
- Observer tests: add test/support/testObserver.ts (createTestObserver,
  createTestObserverCore, createTestIdFactory) and migrate 12 integration tests
  off hand-rolled sqlite/persistence/eventBus/core/api wiring.

Net -523 LOC. No runtime behavior change; absent-vs-undefined optional fields
are preserved (exactOptionalPropertyTypes).

Verify: observer integration 130, observer unit 122, cli observerProviders 19;
biome clean.
