+++ apps/observer/src/commands/session/shared.ts
@@ -1,5 +1,7 @@
 import { randomUUID } from "node:crypto";
 import type {
+  EnsureAgentWorkspaceIntent,
+  HarnessResumeOptions,
   ProviderId,
   ProviderProjectConfig,
   SafeError,
@@ -29,6 +31,7 @@ import { linkAbortSignals, throwIfAborted } from "../cancellation.js";
 
 export { throwIfAborted } from "../cancellation.js";
 
+import { worktreeMissingError } from "../errors.js";
 import type { CommandHandlerContext } from "../queue.js";
 
 export { resolveHarnessProviderOrThrow, resolveTerminalProviderOrThrow } from "../providers.js";
@@ -118,6 +121,139 @@ export function worktreeObservationFromRow(
   }
   if (row.worktree.checks !== undefined) observation.checks = row.worktree.checks;
   return observation;
+}
+
+export type SessionCommandLookupRuntime = {
+  clock?: RuntimeClock | undefined;
+  commandTimeoutMs?: number | undefined;
+  signal?: AbortSignal | undefined;
+  trace?: ProviderMutationTrace | undefined;
+};
+
+export function commandValidationError(input: {
+  code: string;
+  message: string;
+  hint?: string | undefined;
+  projectId?: string | undefined;
+  worktreeId?: string | undefined;
+  sessionId?: string | undefined;
+}): SafeError {
+  const error: SafeError = {
+    tag: "CommandValidationError",
+    code: input.code,
+    message: input.message,
+  };
+  if (input.hint !== undefined) error.hint = input.hint;
+  if (input.projectId !== undefined) error.projectId = input.projectId;
+  if (input.worktreeId !== undefined) error.worktreeId = input.worktreeId;
+  if (input.sessionId !== undefined) error.sessionId = input.sessionId;
+  return error;
+}
+
+export function validateSnapshotRow(row: WorktreeRow | undefined, projectId: string): void {
+  if (row === undefined || row.projectId === projectId) {
+    return;
+  }
+  throw commandValidationError({
+    code: "WORKTREE_PROJECT_MISMATCH",
+    message: "The requested worktree belongs to a different configured project.",
+    projectId,
+    worktreeId: row.id,
+  });
+}
+
+export async function lookupWorktree(input: {
+  providers: ProviderRegistry;
+  projectId: string;
+  worktreeId: string;
+  runtime: SessionCommandLookupRuntime;
+}): Promise<WorktreeObservation> {
+  if (input.providers.worktree.getWorktree === undefined) {
+    throw worktreeMissingError({
+      projectId: input.projectId,
+      worktreeId: input.worktreeId,
+      message: "The requested worktree is not visible to the worktree provider.",
+    });
+  }
+
+  const worktree = await runProviderMutation(
+    {
+      ...input.runtime,
+      operation: `provider.${input.providers.worktree.id}.getWorktree`,
+      fallback: {
+        tag: "WorktreeProviderError",
+        code: "WORKTREE_LOOKUP_FAILED",
+        message: "The worktree provider failed to look up the worktree.",
+        provider: input.providers.worktree.id,
+      },
+    },
+    () =>
+      input.providers.worktree.getWorktree?.({
+        projectId: input.projectId,
+        worktreeId: input.worktreeId,
+      }) as Promise<WorktreeObservation | null>,
+  );
+  if (worktree === null) {
+    throw worktreeMissingError({
+      projectId: input.projectId,
+      worktreeId: input.worktreeId,
+      message: "The requested worktree is not visible to the worktree provider.",
+    });
+  }
+  return worktree;
+}
+
+/**
+ * One workspace-intent envelope for session.create, startAgent, and resumeAgent.
+ * Each caller supplies only the harness fields and resume target it knows about;
+ * absent optional fields stay absent to satisfy exactOptionalPropertyTypes.
+ */
+export function buildEnsureAgentWorkspaceIntent(input: {
+  commandId: string;
+  project: ProviderProjectConfig;
+  worktree: WorktreeObservation;
+  sessionId: string;
+  terminalProvider: string;
+  harnessProvider: string;
+  harness?:
+    | {
+        mode?: "interactive" | "exec" | undefined;
+        profile?: string | undefined;
+        approvalPolicy?: string | undefined;
+        sandboxMode?: string | undefined;
+      }
+    | undefined;
+  layout: string;
+  focus?: boolean | undefined;
+  origin?: EnsureAgentWorkspaceIntent["origin"] | undefined;
+  initialPrompt?: string | undefined;
+  resume?: HarnessResumeOptions | undefined;
+}): EnsureAgentWorkspaceIntent {
+  const intent: EnsureAgentWorkspaceIntent = {
+    type: "session.ensureAgentWorkspace",
+    commandId: input.commandId,
+    terminalProvider: input.terminalProvider,
+    project: input.project,
+    worktree: input.worktree,
+    sessionId: input.sessionId,
+    harness: {
+      provider: input.harnessProvider,
+    },
+    layout: input.layout,
+  };
+  if (input.harness?.mode !== undefined) intent.harness.mode = input.harness.mode;
+  if (input.harness?.profile !== undefined) intent.harness.profile = input.harness.profile;
+  if (input.harness?.approvalPolicy !== undefined) {
+    intent.harness.approvalPolicy = input.harness.approvalPolicy;
+  }
+  if (input.harness?.sandboxMode !== undefined) {
+    intent.harness.sandboxMode = input.harness.sandboxMode;
+  }
+  if (input.resume !== undefined) intent.resume = input.resume;
+  if (input.focus !== undefined) intent.focus = input.focus;
+  if (input.origin !== undefined) intent.origin = input.origin;
+  if (input.initialPrompt !== undefined) intent.initialPrompt = input.initialPrompt;
+  return intent;
 }
 
 /**
