+++ apps/observer/test/integration/reconcile-fake-providers.test.ts
@@ -9,12 +9,8 @@ import {
   FakeWorktreeProvider,
 } from "@station/testing";
 import { describe, expect, it } from "vitest";
-import {
-  createObserverCore,
-  createObserverPersistence,
-  openObserverSqlite,
-  ProviderRegistry,
-} from "../../src/internal";
+import { createObserverCore, ProviderRegistry } from "../../src/internal";
+import { createTestObserverCore } from "../support/testObserver";
 
 const now = "2026-05-20T12:00:00.000Z";
 
@@ -247,23 +243,10 @@ describe("observer reconcile with fake providers", () => {
   });
 
   it("reattaches old branch-derived session bindings to the current path-stable worktree", async () => {
-    const sqlite = openObserverSqlite({ clock: { now: () => new Date(now) } });
-    const persistence = createObserverPersistence({
-      sqlite,
-      clock: { now: () => new Date(now) },
-    });
     const currentWorktreeId = "wt_web_branch_fix_too_path";
     const oldWorktreeId = "wt_web_branch_fix_too_branch";
     const sessionId = "ses_branch_fix_too";
     const worktreePath = "/tmp/station/web/worktrees/branch-fix-too";
-    await persistence.seedSessionTitle({
-      sessionId,
-      projectId: "web",
-      worktreeId: oldWorktreeId,
-      title: "Branch Fix too",
-      createdAt: now,
-      lastSeenAt: now,
-    });
     const providers = new ProviderRegistry({
       worktree: new FakeWorktreeProvider({
         now,
@@ -307,13 +290,18 @@ describe("observer reconcile with fake providers", () => {
         }),
       ],
     });
-    const core = createObserverCore({
+    const { sqlite, persistence, core } = createTestObserverCore({
       config,
       providers,
-      persistence,
-      clock: {
-        now: () => new Date(now),
-      },
+      clock: { now: () => new Date(now) },
+    });
+    await persistence.seedSessionTitle({
+      sessionId,
+      projectId: "web",
+      worktreeId: oldWorktreeId,
+      title: "Branch Fix too",
+      createdAt: now,
+      lastSeenAt: now,
     });
 
     const snapshot = await core.reconcile("old-branch-id-path-reattach");
